add checks to wait for indexes to rebuild
