+++ services/algx-balance-worker.js
@@ -3,7 +3,6 @@ const bullmq = require('bullmq');
 const handleAlgxBalanceJob =
   require('./algx-balance-worker/handleAlgxBalanceJob');
 
-
 const Worker = bullmq.Worker;
 const process = require('node:process');
 // const algosdk = require('algosdk');
