+++ services/asset-worker.js
@@ -4,6 +4,7 @@ const Worker = bullmq.Worker;
 const withSchemaCheck = require('../src/schema/with-db-schema-check');
 const convertQueueURL = require('../src/convert-queue-url');
 const withQueueSchemaCheck = require('../src/schema/with-queue-schema-check');
+const {waitForViewBuildingSimple} = require('./waitForViewBuilding');
 
 const initOrGetIndexer = require('../src/get-indexer');
 
@@ -30,6 +31,7 @@ module.exports = ({queues, databases}) =>{
 
   const assetsWorker = new Worker(convertQueueURL('assets'), async job=>{
     console.log('got assets job ', {assetId: job.data.assetId});
+    await waitForViewBuildingSimple();
     withQueueSchemaCheck('asset', job.data);
     const assetInCouch = await getAssetInCouch(assetDB, job.data.assetId);
     if (assetInCouch) {
