+++ services/trade-history-worker.js
@@ -5,6 +5,7 @@ const withSchemaCheck = require('../src/schema/with-db-schema-check');
 const convertQueueURL = require('../src/convert-queue-url');
 const initOrGetIndexer = require('../src/get-indexer');
 const withQueueSchemaCheck = require('../src/schema/with-queue-schema-check');
+const {waitForViewBuildingSimple} = require('./waitForViewBuilding');
 
 module.exports = ({queues, databases}) =>{
   const blockDB = databases.blocks;
@@ -18,6 +19,8 @@ module.exports = ({queues, databases}) =>{
   const tradeHistoryWorker = new Worker(convertQueueURL('tradeHistory'), async job=>{
     const blockId = job.data.block;
     console.log('received block: ' + blockId);
+    await waitForViewBuildingSimple();
+
     withQueueSchemaCheck('tradeHistory', job.data);
 
     // 1. Get valid escrows from trade history
